Word Cloud #2
Hey guys I looked at the word cloud we had and it's awesome, I was just curious of what it would look like if I lowered the minimum frequency and cut out some of the more expected words like "simul," "model," and "system" to see if more or what would pop up. This is when the min. frequency is set to 50, and those three words I mentioned are taken out. I increased the size of the viewing window to make sure none of the words it wanted to include were cut out. I also included the code for it, but it is literally just what Julia put up just with frequency = 50 not 100 and those three words added to the stop words list.